requirements.txt
fair-esm~=2.0.0
peft~=0.13.2
numpy~=1.26.4
matplotlib~=3.8.0
sckit-learn~=1.5.2
biopython~=1.84